requirements.txt
python-dotenv
requests
gradio
googletrans-py==4.0.0